requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
# tensorflow-gpu>=2.12.0  # Removed - use tensorflow instead

# Computer Vision Libraries
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0
# face_recognition>=1.3.0  # Removed - requires dlib which is problematic on Windows
# dlib>=19.24.0  # Removed - requires CMake and Visual C++
mediapipe>=0.10.0
# insightface>=0.7.0  # Removed - requires Visual C++ build tools

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.14.2

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Progress and Logging
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Web Applications
streamlit>=1.28.0
gradio>=3.40.0
flask>=2.3.0

# Utilities
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0

# Advanced Models
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# Testing and Code Quality
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 